Migrate victory-histogram to TypeScript
