+++ packages/victory-histogram/src/helper-methods.ts
