+++ packages/victory-histogram/src/index.ts
